% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertSBtoNormal}
\alias{convertSBtoNormal}
\title{Convert the vector v into the vector theta/phi using the stick-breaking formula}
\usage{
convertSBtoNormal(v)
}
\arguments{
\item{v}{vector of probabilities (these do not have to sum to 1)}
}
\value{
this function returns a vector res with probabilities that sum to one
}
\description{
This function converts the input vector v into another vector of probabilities 
that sum to one based on the stick-breaking equation
}
