% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_gibbs_functions.R
\name{get.logl}
\alias{get.logl}
\title{Calculate the log-likelihood}
\usage{
get.logl(y, omega, nspp, nloc, xmat, betas, cs, alpha)
}
\arguments{
\item{y}{L x S matrix, with the presence-absence data}

\item{omega}{L x S matrix with the underlying latent variables omega}

\item{nspp}{overall number of species (S)}

\item{nloc}{overall number of locations (L)}

\item{xmat}{L x P design matrix containing the covariates (columns) for each location (rows). 
Notice that this matrix does not contain a column of 1's for the intercept}

\item{betas}{P x KS matrix containing slope parameters for each covariate (row) and each group (column)}

\item{cs}{vector of length S wth group assignment for each species}

\item{alpha}{vector of length S with intercept for each species}
}
\value{
log-likelihood of probit regression models
}
\description{
This function calculates the log-likelihood based on the current parameter estimates
}
