% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getql}
\alias{getql}
\title{Summarize data for SBM according to the different location and species groups}
\usage{
getql(z, w, dat, ngrloc, ngrspp)
}
\arguments{
\item{z}{integer vector with cluster assignments of locations}

\item{w}{integer vector with cluster assignments of species}

\item{dat}{this matrix has L rows (e.g., locations) and S columns (e.g., species)
and contains the presence-absence data}

\item{ngrloc}{number of location groups (KL)}

\item{ngrspp}{number of species groups (KS)}
}
\value{
this function returns a list with two KL x KS matrices: nql1 and nql0
}
\description{
This function determines how many observations are in each location group 
and species group for two cases: 
- when dat(i,j)=1: these results are stored in matrix nql1
- when dat(i,j)=0: these results are stored in matrix nql0
}
