% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_main.R
\name{gibbs.SAM}
\alias{gibbs.SAM}
\title{Main function of the Species Archetype Model (SAM)}
\usage{
gibbs.SAM(y, xmat, ngroups, ngibbs, burnin)
}
\arguments{
\item{y}{this matrix has L rows (locations) and S columns (species) and contains the presence-absence data}

\item{xmat}{L X P design matrix containing the covariates (columns) for each
location (rows). Notice that this matrix does not contain a column of 1's for the intercept}

\item{ngroups}{maximum number of species groups (K)}

\item{ngibbs}{number of Gibbs sampler iterations}

\item{burnin}{number of iterations to discard as part of burn-in phase}
}
\value{
this function returns a list containing several matrices, all of which have
        ngibbs-burnin rows, containing samples from the posterior distribution of:
              \itemize{
                 \item logl:  log-likelihood for each iteration
                 \item theta: probability associated with each species group
                 \item betas: slope parameters for each group
                 \item cs:    cluster assignment of each species
                 \item alpha: intercept of each species
                 \item gamma: TSB prior parameter
              }
}
\description{
Runs the Gibbs sampler and returns samples from the posterior distribution
}
