% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_main.R
\name{mixture.gibbs.main.func}
\alias{mixture.gibbs.main.func}
\title{Main function of the Mixture Model}
\usage{
mixture.gibbs.main.func(dat, ngroup, ngibbs, burnin)
}
\arguments{
\item{dat}{this matrix has L rows (locations) and S columns (species)
and contains the presence-absence data}

\item{ngroup}{maximum number of location groups (K)}

\item{ngibbs}{number of Gibbs sampler iterations}

\item{burnin}{number of iterations to discard as burn-in}
}
\value{
this function returns a list containing several matrices.
        These matrices have ngibbs-burnin rows and contain samples from the posterior distribution for:
        \itemize{
           \item phi:   probability of observing each species in each group
           \item theta: probability of each location group
           \item logl:  log-likelihood
           \item z:     cluster assignment of each location
           \item gamma: TSB prior parameter
        }
}
\description{
Runs the Gibbs sampler and returns samples from the posterior distribution
}
