% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_main.R
\name{mixture.gibbs}
\alias{mixture.gibbs}
\title{Runs the Gibbs sampler and returns samples from the posterior distribution}
\usage{
mixture.gibbs(dat, ngroup, nl, ngibbs, burnin, a.prior, b.prior)
}
\arguments{
\item{dat}{this matrix has L rows (locations) and S columns (species)
and contains the presence-absence data (i.e., number of times 
a given species was observed at a given location)}

\item{ngroup}{maximum number of location groups (K)}

\item{nl}{this vector has L elements (locations) and contains the number of observation 
opportunities at each location}

\item{ngibbs}{number of Gibbs sampler iterations}

\item{burnin}{number of iterations to discard as burn-in}

\item{a.prior}{'a' parameter for prior beta distribution}

\item{b.prior}{'b' parameter for prior beta distribution}
}
\value{
this function returns a list containing several matrices.
        These matrices have ngibbs-burnin rows and contain samples from the posterior distribution for:
        \itemize{
           \item phi:   probability of observing each species in each group
           \item theta: probability of each location group
           \item logl:  log-likelihood
           \item z:     cluster assignment of each location
           \item gamma: TSB prior parameter
        }
}
\description{
Runs the Gibbs sampler and returns samples from the posterior distribution
}
