% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ncs}
\alias{ncs}
\title{Summarize the data for mixture model}
\usage{
ncs(dat, z, nspp, nloc, ngroup)
}
\arguments{
\item{dat}{matrix with L rows (e.g., locations) and S columns (e.g., species),
containing the presence-absence data}

\item{z}{vector of size L with cluster assignment for each location}

\item{nspp}{number of species (S)}

\item{nloc}{number of locations (L)}

\item{ngroup}{maximum number of groups (K)}
}
\value{
this function returns a list containing two matrices of size K x S: res1 and res0
}
\description{
This function summarizes the data by calculating the number of locations for each 
location group and species for which dat(i,j)=1 (stored in res1) and dat(i,j)=0 (stored in res0)
}
