% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_gibbs_functions.R
\name{sample.betas}
\alias{sample.betas}
\title{Sample betas}
\usage{
sample.betas(ngroups, cs, nparam, xtx, t.xmat, alpha, nloc, nspp, omega)
}
\arguments{
\item{ngroups}{maximum number of species groups (KS)}

\item{cs}{vector of size S with group assignment for each species}

\item{nparam}{overall number of covariates (P)}

\item{xtx}{equal to t(X)*X}

\item{t.xmat}{equal to tranpose of the design matrix t(X)}

\item{alpha}{vector of size S with intercept for each species}

\item{nloc}{overall number of locations (L)}

\item{nspp}{overall number of species (S)}

\item{omega}{a matrix L x S containing the underlying latent variables omega}
}
\value{
this function returns a P x KS matrix with regression slope parameters betas
}
\description{
This function generates samples of betas, the regression slope parameters
}
