% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_gibbs_functions.R
\name{sample.cs}
\alias{sample.cs}
\title{Sample cs}
\usage{
sample.cs(ngroups, omega, xmat, alpha, betas, theta, nspp, nloc,
  InvSigma.precalc, Sigma.precalc, lds, cs, nparam)
}
\arguments{
\item{ngroups}{maximum number of species groups (KS)}

\item{omega}{a matrix L x S containing the underlying latent variables omega}

\item{xmat}{the design matrix of dimension L x P}

\item{alpha}{vector of size S with intercept for each species}

\item{betas}{P x KS matrix containing slope parameters for each covariate (row) and each group (column)}

\item{theta}{vector of size KS with the probability of each species group}

\item{nspp}{overall number of species (S)}

\item{nloc}{overall number of locations (L)}

\item{Sigma.precalc}{covariance matrix for new group}

\item{lds}{log of determinant of covariance matrix for new group}

\item{cs}{vector of size S with group assignment for each species}

\item{nparam}{overall number of covariates (P)}

\item{invSigma.precalc}{precision matrix for new group}

\item{return}{a vector of size S with the group assignment variable cs}
}
\description{
This function generates samples of the group assignment variable cs
}
