% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_gibbs_functions.R,
%   R/mixture_gibbs_functions.R
\name{sample.gamma}
\alias{sample.gamma}
\alias{sample.gamma}
\title{Sample gamma}
\usage{
sample.gamma(v, ngroup, gamma.possib)

sample.gamma(v, ngroup, gamma.possib)
}
\arguments{
\item{v}{vector of size KS with the stick-breaking parameters}

\item{ngroup}{maximum number of location groups (K)}

\item{gamma.possib}{set of possible values for gamma}

\item{ngroups}{maximum number of species groups (KS)}

\item{return}{a sample of gamma}

\item{v}{vector of length L with probabilities}

\item{gamma.possib}{vector of possible gamma parameter values}

\item{return}{this function returns a real number corresponding to gamma}
}
\description{
This function generates a sample of the truncated stick breaking prior parameter gamma

This function samples the truncated stick breaking (TSB) prior parameter gamma
}
