% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_gibbs_functions.R
\name{sample.gamma.sam}
\alias{sample.gamma.sam}
\title{Sample gamma}
\usage{
sample.gamma.sam(v, ngroups, gamma.possib)
}
\arguments{
\item{v}{vector of size KS with the stick-breaking parameters}

\item{ngroups}{maximum number of species groups (KS)}

\item{gamma.possib}{set of possible values for gamma}
}
\value{
a sample of gamma
}
\description{
This function generates a sample of the truncated stick breaking prior parameter gamma
}
