% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_gibbs_functions.R
\name{sample.gamma.u}
\alias{sample.gamma.u}
\title{Sample gamma.u}
\usage{
sample.gamma.u(uk, gamma.possib, ngroup.spp)
}
\arguments{
\item{uk}{vector of size KS with probabilities}

\item{gamma.possib}{this is a vector containing the possible values that gamma.u can take}

\item{ngroup.spp}{maximum number of species groups (KS)}

\item{return}{this function returns a real number (gamma.u)}
}
\description{
Sample the TSB prior parameter for species groups (gamma.u) from its full conditional distribution
}
