% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_gibbs_functions.R
\name{sample.gamma.v}
\alias{sample.gamma.v}
\title{Sample gamma.v}
\usage{
sample.gamma.v(vk, gamma.possib, ngroup.loc)
}
\arguments{
\item{vk}{vector of size KL with probabilities}

\item{gamma.possib}{vector containing the possible values that gamma.u can take}

\item{ngroup.loc}{maximum number of location groups (KL)}
}
\value{
this function returns a real number (gamma.v)
}
\description{
Sample the TSB prior parameter for location groups (gamma.v) from its full conditional distribution
}
