% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_gibbs_functions.R
\name{sample.omega}
\alias{sample.omega}
\title{Sample omega}
\usage{
sample.omega(y, nspp, nloc, xmat, betas, cs, alpha)
}
\arguments{
\item{y}{L x S matrix, containing the presence-absence data}

\item{nspp}{overall number of species S}

\item{nloc}{overall number of locations L}

\item{xmat}{L x P matrix containing the covariates (columns) for each location (rows). 
Notice that this matrix does not contain a column of 1's for the intercept}

\item{betas}{P x KS matrix containing slope parameters for each covariate (row) for each group (column)}

\item{cs}{group assignment for each species}

\item{alpha}{intercept for each species}

\item{n}{number of samples to generate}

\item{return}{this function returns a L x S matrix containing the underlying latent variables omega}
}
\description{
This function generates samples of omega (the underlying latent continuous variable)
}
