% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_gibbs_functions.R
\name{sample.phi}
\alias{sample.phi}
\title{Sample phi and uk parameters}
\usage{
sample.phi(ngroup.spp, gamma.u, burnin, gibbs.step, phi, psi, w)
}
\arguments{
\item{ngroup.spp}{maximum number of groups for species (KS)}

\item{gamma.u}{the truncated stick-breaking prior parameter for
species groups. This value should be between 0 and 1, and
small values enforce more parsimonius results (i.e., fewer groups)}

\item{burnin}{number of MCMC samples that are going to be thrown out as
part of the burn-in phrase}

\item{gibbs.step}{current iteration of the gibbs sampler}

\item{phi}{vector of size KS containing the current estimate of phi (i.e., probability of each species group)}

\item{psi}{matrix of size KL x KS containing the current estimate  of psi}

\item{w}{vector of length S, storing the current membership of each species}
}
\value{
this function returns a list with 4 items (phi, uk, w, psi)
}
\description{
Sample uk parameters from their full conditional distributions and
then calculate the implied phi parameters
}
