% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_gibbs_functions.R
\name{sample.psi}
\alias{sample.psi}
\title{Sample psi parameters}
\usage{
sample.psi(z, w, dat, ngroup.loc, ngroup.spp)
}
\arguments{
\item{z}{vector of length L, storing the current membership of each location}

\item{w}{vector of length S, storing the current membership of each species}

\item{dat}{L x S matrix containing the presence-absence data}

\item{ngroup.loc}{maximum number of location groups (KL)}

\item{ngroup.spp}{maximum number of species groups (KS)}

\item{return}{a KL x KS matrix of psi parameters}
}
\description{
Sample the presence probability for each location and species group (psi) from 
it's full conditional distribution
}
