% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_gibbs_functions.R
\name{sample.theta.sam}
\alias{sample.theta.sam}
\title{Sample theta}
\usage{
sample.theta.sam(cs, ngroups, gamma, burnin, gibbs.step, betas, theta)
}
\arguments{
\item{cs}{vector of size S with group assignment for each species}

\item{ngroups}{maximum number of species groups (KS)}

\item{gamma}{the parameter for the stick-breaking prior}

\item{burnin}{number of iterations to discard as part of burn-in phase}

\item{gibbs.step}{current iteration of the gibbs sampler}

\item{betas}{P x KS matrix containing slope parameters for each covariate (row) and each group (column)}

\item{theta}{vector of size KS with the probability of each species group}
}
\value{
a list of 4 items (theta,betas,cs,v)
}
\description{
This function generates a sample of the vector of probabilities associated with each species group
}
