% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_gibbs_functions.R
\name{sample.w}
\alias{sample.w}
\title{Sample w}
\usage{
sample.w(lphi, dat, dat1m, lpsi, l1mpsi, ngroup.spp, ngroup.loc, nloc,
  nspp, w, z)
}
\arguments{
\item{lphi}{equal to log(phi)}

\item{dat}{L x S matrix, containing the presence-absence data}

\item{dat1m}{L x S matrix, calculated as 1-dat}

\item{lpsi}{equal to log(psi)}

\item{l1mpsi}{equal to log(1-psi)}

\item{ngroup.spp}{thismaximum number of species groups (KS)}

\item{ngroup.loc}{maximum number of location groups (KL)}

\item{nloc}{total number of locations (L)}

\item{nspp}{total number of species (S)}

\item{w}{vector of length S, storing the current membership of each species}

\item{z}{vector of length L, storing the current membership of each locations}
}
\value{
this function returns a vector of length S containing w
}
\description{
Sample the vector w containing the membership of each species
}
