% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_gibbs_functions.R
\name{sample.z}
\alias{sample.z}
\title{Sample z}
\usage{
sample.z(ltheta, dat, dat1m, lpsi, l1mpsi, ngroup.loc, ngroup.spp, nloc, nspp,
  w, z)
}
\arguments{
\item{ltheta}{equal to log(theta)}

\item{dat}{L x S matrix, containing the presence-absence data}

\item{dat1m}{L x S matrix, calculated as 1-dat}

\item{lpsi}{equal to log(psi)}

\item{l1mpsi}{equal to log(1-psi)}

\item{ngroup.loc}{maximum number of location groups (KL)}

\item{ngroup.spp}{maximum number of species groups (KS)}

\item{nloc}{total number of locations (L)}

\item{nspp}{total number of species (S)}

\item{w}{vector of length S, storing the current membership of each species}

\item{z}{vector of length L, storing the current membership of each locations}

\item{return}{this function returns a vector of length L containing z}
}
\description{
Sample the vector z containing the membership of each location
}
