% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBM_main.R
\name{SBM}
\alias{SBM}
\title{Main function of the Stochastic Block Model (SBM)}
\usage{
SBM(dat, ngroup.loc, ngroup.spp, ngibbs, burnin)
}
\arguments{
\item{dat}{matrix has L rows (e.g., locations) and S columns (e.g., species),
containing the presence-absence data}

\item{ngroup.loc}{maximum number of location groups (KL)}

\item{ngroup.spp}{maximum number of species groups (KS)}

\item{ngibbs}{number of Gibbs sampler iterations}

\item{burnin}{number of iterations to discard as part of burn-in phase}

\item{return}{this function returns samples from the posterior distribution in the form of a 
list containing several matrices. These matrices have ngibbs-burnin rows and 
contain the posterior distribution for:
- theta: probability of each location group
- phi:   probability of each species group
- llk:   log-likelihood for each iteration
- psi:   presence probability for each location group and species group                                
- z:     cluster assignment of each location
- w:     cluster assignment of each species
- gamma: TSB prior parameters (one for location groups and the other for species groups)}
}
\description{
Runs the Gibbs sampler and returns samples from the posterior distribution
}
