% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_gibbs_functions.R
\name{update.theta}
\alias{update.theta}
\title{Samples theta and v parameters}
\usage{
\method{update}{theta}(z, ngroup, gamma1, burnin, gibbs.step, theta, phi)
}
\arguments{
\item{z}{vector of size L with cluster assignment of each location}

\item{ngroup}{maximum number of location groups (K)}

\item{gamma1}{this is the truncated stick-breaking prior parameter for the 
number of location groups. This value should be between 0 and 1, and
small values enforce more parsimonius results (i.e., fewer groups)}

\item{burnin}{number of iterations to drop as part of burn-in phase}

\item{gibbs.step}{current iteration of the gibbs sampler}

\item{theta}{vector of length K with the probability of each location group}

\item{phi}{K x S matrix with the probability of observing each species in each group}
}
\value{
this function returns a list of 4 items (theta, z, v, and phi)
}
\description{
This function samples the v parameters, which are then used to calculate the theta parameters
}
