% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_gibbs_functions.R
\name{update.z}
\alias{update.z}
\title{Samples z}
\usage{
\method{update}{z}(
  dat,
  nl,
  n.minus.y,
  phi,
  theta,
  ngroup,
  nloc,
  nspp,
  z,
  a.prior,
  b.prior,
  constant
)
}
\arguments{
\item{dat}{this matrix has L rows (locations) and S columns (species)
and contains the presence-absence data (i.e., number of times 
a given species was observed at a given location)}

\item{nl}{this vector has L elements (locations) and contains the number of observation 
opportunities at each location}

\item{n.minus.y}{matrix with L rows (e.g., locations) and S columns (e.g., species),
calculated as matrix(nl,L,S)-dat}

\item{phi}{K x S matrix with the probability of observing each species in each group}

\item{theta}{vector of length K with the probability of each location group}

\item{ngroup}{maximum number of location groups (K)}

\item{nloc}{number of locations (L)}

\item{nspp}{number of species (S)}

\item{z}{vector of size L containing the current cluster assignment for each location}

\item{a.prior}{'a' parameter for prior beta distribution}

\item{b.prior}{'b' parameter for prior beta distribution}

\item{constant}{constant term used to calculate probability for a new group}
}
\value{
this function returns a vector of size L with the cluster assignment of each location
}
\description{
This function samples the cluster assignment for each location (z)
}
