% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_main.R
\name{gibbs.SAM}
\alias{gibbs.SAM}
\title{Main function of the Species Archetype Model (SAM)}
\usage{
gibbs.SAM(y, xmat, ngroups, ngibbs, burnin)
}
\arguments{
\item{y}{L X S matrix, containing the presence-absence data}

\item{xmat}{L X P design matrix containing the covariates (columns) for each
location (rows). Notice that this matrix does not contain a column of 1's for the intercept}

\item{ngroups}{maximum number of species groups (KS)}

\item{ngibbs}{number of Gibbs sampler iterations}

\item{burnin}{number of iterations to discard as part of burn-in phase}

\item{return}{this function returns a list containing several matrices, all of which have 
ngibbs-burnin rows, containing samples from the posterior distribution of:
- theta: probability associated with each species group
- logl:  log-likelihood for each iteration
- betas: slope parameters for each group
- cs:    cluster assignment of each species
- alpha: intercept of each species
- gamma: TSB prior parameter}
}
\description{
Runs the Gibbs sampler and returns samples from the posterior distribution
}
