% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_gibbs_functions.R, R/SBM_gibbs_functions.R
\name{sample.theta}
\alias{sample.theta}
\alias{sample.theta}
\title{Sample theta}
\usage{
sample.theta(ngroup.loc, gamma.v, burnin, gibbs.step, theta, psi, z)

sample.theta(ngroup.loc, gamma.v, burnin, gibbs.step, theta, psi, z)
}
\arguments{
\item{ngroup.loc}{maximum number of groups for locations (KL)}

\item{gamma.v}{truncated stick-breaking prior parameter for the 
location groups. This value should be between 0 and 1, and
small values enforce more parsimonius results (i.e., fewer groups)}

\item{burnin}{number of iterations to discard as part of burn-in phase}

\item{gibbs.step}{current iteration of the gibbs sampler}

\item{theta}{vector of size KS with the probability of each species group}

\item{psi}{KL x KS matrix  containing the current estimate of psi}

\item{z}{this is a vector of length L, storing the current membership of each location}

\item{cs}{vector of size S with group assignment for each species}

\item{ngroups}{maximum number of species groups (KS)}

\item{gamma}{the parameter for the stick-breaking prior}

\item{betas}{P x KS matrix containing slope parameters for each covariate (row) and each group (column)}

\item{return}{a list of 4 items (theta,betas,cs,v)}

\item{burnin}{number of MCMC samples that are going to be thrown out as 
part of the burn-in phrase}

\item{gibbs.step}{current iteration of the gibbs sampler}

\item{theta}{vector of size KL containing the current estimate of theta (i.e., probability of each location group)}

\item{return}{this function returns a list containing 4 items (theta, vk, psi, and z)}
}
\description{
This function generates a sample of the vector of probabilities associated with each species group

Sample vk parameters from their full conditional distributions and convert 
these vk parameters into theta parameters
}
