% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_gibbs_functions.R
\name{update.z}
\alias{update.z}
\title{Samples z}
\usage{
\method{update}{z}(dat, one.minus.dat, phi, theta, ngroup, nloc, nspp, z)
}
\arguments{
\item{dat}{matrix with L rows (e.g., locations) and S columns (e.g., species),
containing the presence-absence data}

\item{one.minus.dat}{matrix with L rows (e.g., locations) and S columns (e.g., species),
calculated as 1-dat}

\item{phi}{K x S matrix with the probability of observing each species in each group}

\item{theta}{vector of length K with the probability of each location group}

\item{ngroup}{maximum number of location groups (K)}

\item{nloc}{number of locations (L)}

\item{nspp}{number of species (S)}

\item{z}{vector of size L containing the current cluster assignment for each location}

\item{return}{this function returns a vector of size L with the cluster assignment of each location}
}
\description{
This function samples the cluster assignment for each location (z)
}
